(#206) unificando usuarios em heranca
