+++ app/model/professor.py
@@ -1,40 +1,20 @@
+from .user import User
 from ..ext.database import db
-from passlib.hash import pbkdf2_sha256
 
 
-class Professor(db.Model):
+class Professor(User):
     __tablename__ = "professor"
 
     id_professor = db.Column(db.Integer, primary_key=True, autoincrement=True)
     reg_professor = db.Column(db.BigInteger, nullable=True, unique=True)
-
-    name = db.Column(db.String(255), nullable=False)
-    email = db.Column(db.String(255), nullable=True, unique=True)
-    __password_hash = db.Column("password", db.String(255), nullable=True)
-
     disciplines = db.relationship(
         "Discipline", secondary="professor_discipline", lazy="dynamic"
     )
-
     posts = db.relationship("Post", back_populates="professor")
 
     @property
     def reg(self):
         return self.reg_professor
-
-    def is_professor(self):
-        return True
-
-    @property
-    def password(self):
-        raise AttributeError("password: write-only field")
-
-    @password.setter
-    def password(self, password):
-        self.__password_hash = pbkdf2_sha256.hash(password)
-
-    def verify_password(self, password):
-        return pbkdf2_sha256.verify(password, self.__password_hash)
 
     @property
     def rating(self):
@@ -45,11 +25,6 @@ class Professor(db.Model):
         for post in self.posts:
             sum += post.rating
         return sum / len(self.posts)
-
-    def clean_credentials(self):
-        self.__password_hash = None
-        self.email = None
-        self.reg_professor = None
 
     @property
     def didactic(self):
@@ -90,3 +65,11 @@ class Professor(db.Model):
         for post in self.posts:
             sum += post.disponibility
         return sum / len(self.posts)
+
+    def is_professor(self):
+        return True
+
+    def clean_credentials(self):
+        self.__password_hash = None
+        self.email = None
+        self.reg_professor = None
