+++ app/model/user.py
@@ -1,13 +1,36 @@
-from ..services import user_services
+from ..ext.database import db
+from passlib.hash import pbkdf2_sha256
 
 
-class User:
-    def __init__(self, email, password):
-        self.email = email
-        self.password = password
+class User(db.Model):
+    __abstract__ = True
+
+    name = db.Column(db.String(255), nullable=False)
+    email = db.Column(db.String(100), nullable=True, unique=True)
+    __email_verified = db.Column(
+        "email_verified", db.Boolean, nullable=False, default=False
+    )
+    __password_hash = db.Column("password", db.String(255))
+
+    @property
+    def password(self):
+        raise AttributeError("password: write-only field")
+
+    @password.setter
+    def password(self, password):
+        self.__password_hash = pbkdf2_sha256.hash(password)
 
     def is_professor(self):
-        return user_services.is_professor(email=self.email)
+        pass
 
-    def get(self):
-        return user_services.get(email=self.email)
+    def is_email_verified(self):
+        return self.__email_verified
+
+    def active_user(self):
+        self.__email_verified = True
+
+    def verify_password(self, password):
+        return pbkdf2_sha256.verify(password, self.__password_hash)
+
+    def save_changes(self):
+        db.session.commit()
